feat: extract requirement data classes
